Updates for fid search area.
git-tfs-id: [https://devops.cyberoptics.com/TPC1]$/CyberStitch;C33024

+++ Main/src/logic/ManagedPanelAlignment/ManagedPanelAlignment.cpp
@@ -32,7 +32,7 @@ namespace PanelAlignM {
 
 	int ManagedPanelAlignment::GetNumberOfFidsProcessed()
 	{
-		return _pAligner->GetLastProcessedFids()->size();
+		return (int)_pAligner->GetLastProcessedFids()->size();
 	}
 
 	ManagedFidInfo^ ManagedPanelAlignment::GetFidAtIndex(unsigned int index)
